% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBGapp.R
\docType{package}
\name{MBGapp-package}
\alias{MBGapp}
\alias{MBGapp-package}
\title{MBGapp: Interactive Shiny Application for Model-Based Geostatistics}
\description{
The MBGapp package provides a Shiny web application for teaching and applied
geostatistical analysis. Users can explore spatial data interactively, assess
spatial correlation through the empirical variogram, fit model-based
geostatistical models for continuous, prevalence and count outcomes, generate
spatial predictions, and download reports.
}
\details{
The application is launched with \code{\link{run_app}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/olatunjijohnson/MBGapp}
  \item Report bugs at \url{https://github.com/olatunjijohnson/MBGapp/issues}
}

}
\author{
\strong{Maintainer}: Olatunji Johnson \email{olatunjijohnson21111@gmail.com} (\href{https://orcid.org/0000-0002-4080-0999}{ORCID})

Authors:
\itemize{
  \item Claudio Fronterre
  \item Emanuele Giorgi
}

}
\keyword{internal}
